refactoring: MemberController 에서 AuthController 변경
